refactor: break Registration out to smaller parts
