+++ src/App.tsx
@@ -3,46 +3,39 @@ import { Router } from '@reach/router'
 import RestrictedRoute from 'views/restricted'
 import Loader from 'components/Loader'
 
-const Login = React.lazy(() => import(`views/Login/Login`))
-const Start = React.lazy(() => import(`views/Start/Start`))
+const Login = React.lazy(() => import('views/Login/Login'))
+const Start = React.lazy(() => import('views/Start/Start'))
 const Profile = React.lazy(() => import('views/Profile/Profile'))
-
 const RegistrationCompleted = React.lazy(() =>
   import('views/CreateProfile/RegistrationCompleted')
 )
-
 const CreateProfile = React.lazy(() =>
   import('views/CreateProfile/CreateProfile')
 )
 const ChooseProfession = React.lazy(() =>
-  import(`views/CreateProfile/ChooseProfession`)
+  import('views/CreateProfile/ChooseProfession')
 )
 const MatchCompetences = React.lazy(() =>
-  import('views/CreateProfile/MatchSkills')
+  import('views/CreateProfile/AddSkills')
 )
 const AddEducation = React.lazy(() =>
   import('views/CreateProfile/AddEducation')
 )
 const AddContactInformation = React.lazy(() =>
-  import(`views/CreateProfile/AddContactInformation`)
+  import('views/CreateProfile/AddContactInformation')
 )
-
 const PreviousOccupationExperience = React.lazy(() =>
   import('views/CreateProfile/PreviousOccupationExperience')
 )
-
 const SaveCV = React.lazy(() => import('views/CreateProfile/SaveCV'))
-
 const WorkExperiences = React.lazy(() =>
   import('views/CreateProfile/WorkExperiences')
 )
-
 const UploadImage = React.lazy(() => import('views/CreateProfile/UploadImage'))
-
-const WhoAmI = React.lazy(() => import(`views/CreateProfile/WhoAmI`))
-const AddTraits = React.lazy(() => import(`views/CreateProfile/AddTraits`))
-const Home = React.lazy(() => import(`views/Profile/Home`))
-const Timeline = React.lazy(() => import(`views/Profile/Timeline`))
+const WhoAmI = React.lazy(() => import('views/CreateProfile/WhoAmI'))
+const AddTraits = React.lazy(() => import('views/CreateProfile/AddTraits'))
+const Home = React.lazy(() => import('views/Profile/Home'))
+const Timeline = React.lazy(() => import('views/Profile/Timeline'))
 
 function App() {
   return (
@@ -52,17 +45,17 @@ function App() {
           <Start path="/" />
           <Login path="/login" />
           <CreateProfile path="/skapa-cv">
-            <AddEducation path="/utbildning" />
-            <ChooseProfession path="/" />
-            <RegistrationCompleted path="/grattis" />
-            <MatchCompetences path="/kompetenser" />
-            <WhoAmI path="/beskriv-dig" />
-            <AddTraits path="/egenskaper" />
-            <UploadImage path="/profilbild" />
-            <PreviousOccupationExperience path="/erfarenheter" />
-            <WorkExperiences path="/erfarenheter/tidigare-erfarenheter" />
             <AddContactInformation path="/kontakt" />
+            <AddEducation path="/utbildning" />
+            <AddTraits path="/egenskaper" />
+            <ChooseProfession path="/" />
+            <MatchCompetences path="/kompetenser" />
+            <PreviousOccupationExperience path="/erfarenheter" />
+            <RegistrationCompleted path="/grattis" />
             <SaveCV path="/spara-cv" />
+            <UploadImage path="/profilbild" />
+            <WhoAmI path="/beskriv-dig" />
+            <WorkExperiences path="/erfarenheter/tidigare-erfarenheter" />
           </CreateProfile>
 
           <Profile path="/profil">
