+++ src/views/CreateProfile/__tests__/AddTraits.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import AddTraits, { ADD_TRAIT } from '../AddTraits'
-import { GET_ONTOLOGY_CONCEPTS } from '../ChooseProfession'
+import AddTraits, { ADD_TRAIT } from 'views/partials/Traits'
+import { GET_ONTOLOGY_CONCEPTS } from 'views/partials/Profession'
 import { render } from 'utils/test-utils'
 import { waitForElement, fireEvent } from 'react-testing-library'
 import { OntologyType } from 'generated/myskills.d'
